scalar_prod method of the Vector class truncated the result if it did not fit into 32 bit, even though it returns a 64-bit value

+++ shapemerge2d/vector.h
@@ -22,7 +22,7 @@ namespace shapemerge2d
 		bool operator==(const Vector& o) const;
 		int64_t scalarprod(const Vector& o) const
 		{
-			return x*o.x+y*o.y;
+		        return int64_t(x)*o.x+int64_t(y)*o.y;
 		}
 		std::string __repr__() const
 		{
